Continue rewrite, caching is gone for now
