+++ src-tauri/src/plugins/ugg.rs
@@ -1,5 +1,6 @@
 pub mod constants;
 pub mod structs;
+mod json;
 mod shards;
 mod rates;
 mod runes;
