+++ src-tauri/src/plugins/ugg/items.rs
@@ -1,6 +1,6 @@
 use crate::shared::data_dragon::structs::DataDragon;
 
-use super::{structs::{self, ItemsMap, ItemValues}, requests::overview, constants::DATA};
+use super::{structs::{self, ItemsMap, ItemValues}, json::overview, constants::DATA};
 
 impl structs::Data {
